refactor : useFilteredEvents 추가
